style sep jsx file
