+++ react-cors-spa/src/AddProduct.jsx
@@ -1,21 +1,11 @@
 import * as React from 'react';
 import { Button } from "@mui/material";
-import { styled } from '@mui/material/styles';
 import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
 import DialogContent from '@mui/material/DialogContent';
 import DialogContentText from '@mui/material/DialogContentText';
 import DialogTitle from '@mui/material/DialogTitle';
-
-const AddProductButton = styled(Button)({
-    color: 'white',
-    backgroundColor: '#1c1c24',
-    borderColor: 'grey',
-    '&:hover': {
-    backgroundColor: '#24242e',
-    boxShadow: '0 12px 16px 0 rgba(0,0,0,0.24),0 17px 50px 0 rgba(0,0,0,0.19)',
-  },
-  });
+import AddProductButton from './AddProductStyles';
 
 function AddProduct() {
     const [open, setOpen] = React.useState(false);
